requirements.txt
asgi-redis==1.2.0
asgiref==1.1.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.4.2
Django==2.0.6
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack-python==0.5.6
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2018.4
redis==2.10.6
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
zope.interface==4.5.0
